Pull reactions into reportbackItem state object.
This matches how we're structuring posts & reactions in Rogue, and
simplifies our lives so we don't have to touch multiple parts of the
state tree in response to a single request.
