+++ resources/assets/actions/reportback.js
@@ -3,6 +3,8 @@ import { normalizeReportbacksResponse } from "../normalizers";
 import {
   REQUESTED_REPORTBACKS,
   RECEIVED_REPORTBACKS,
+  REACTION_CHANGED,
+  REACTION_COMPLETE,
   STORE_REPORTBACK_PENDING,
   STORE_REPORTBACK_FAILED,
   STORE_REPORTBACK_SUCCESSFUL,
@@ -25,6 +27,20 @@ export function requestedReportbacks(node) {
 // Action: new reportback data received.
 export function receivedReportbacks(page, { reportbacks, reportbackItems, reactions }) {
   return { type: RECEIVED_REPORTBACKS, page, reportbacks, reportbackItems, reactions };
+}
+
+// Action: toggled a reaction
+export function reactionChanged(reportbackItemId, value) {
+  return { type: REACTION_CHANGED, reportbackItemId, value };
+}
+
+// Action: component got a reaction response back.
+export function reactionComplete(reportbackItemId, reactionId) {
+  return {
+    type: REACTION_COMPLETE,
+    reportbackItemId,
+    reactionId,
+  }
 }
 
 // Action: store new user submitted reportback.
@@ -62,6 +78,32 @@ export function addToSubmissionsList(reportback) {
   return {
     type: ADD_TO_SUBMISSIONS_LIST,
     reportback
+  }
+}
+
+// Async Action: user reacted to a photo.
+export function toggleReactionOn(reportbackItemId, termId) {
+  return dispatch => {
+    dispatch(reactionChanged(reportbackItemId, true));
+
+    (new Phoenix).post('reactions', {
+      reportback_item_id: reportbackItemId,
+      term_id: termId,
+    })
+      .then(json => {
+        if (json && json[0] && json[0].created) {
+          dispatch(reactionComplete(reportbackItemId, json[0].kid));
+        }
+      });
+  }
+}
+
+// Async Action: user un-reacted to a photo.
+export function toggleReactionOff(reportbackItemId, reactionId) {
+  return dispatch => {
+    dispatch(reactionChanged(reportbackItemId, false));
+
+    (new Phoenix).delete(`reactions/${reactionId}`);
   }
 }
