+++ resources/assets/normalizers.js
@@ -6,17 +6,18 @@
 export function normalizeReportbacksResponse(data) {
   let reportbacks = {};
   let reportbackItems = {};
-  let reactions = {};
 
   data.forEach((reportback) => {
     reportback.reportback_items = reportback.reportback_items.data.map((item) => {
       const currentUser = item.kudos.data[0] ? item.kudos.data[0].current_user : false;
 
-      reactions[item.id] = {
+      item.reaction = {
         id: currentUser ? currentUser.kudos_id : null,
+        reacted: !!(currentUser && currentUser.kudos_id),
         total: item.kudos.data[0] ? item.kudos.data[0].term.total : 0,
         termId: item.kudos.data[0] ? item.kudos.data[0].term.id : '1274', // This is a hardcoded default because phoenix-ashes is bugged.
       };
+      delete item.kudos;
 
       reportbackItems[item.id] = item;
       return item.id;
@@ -26,5 +27,5 @@ export function normalizeReportbacksResponse(data) {
     reportbacks[reportback.id] = reportback;
   });
 
-  return {reportbacks, reportbackItems, reactions};
+  return {reportbacks, reportbackItems};
 }
