+++ resources/assets/reducers/index.js
@@ -3,5 +3,4 @@ export reportbacks from './reportbacks';
 export submissions from './submissions';
 export blocks from './blocks';
 export user from './user';
-export reactions from './reactions';
 export signups from './signups';
